\name{FREEcvOutput}
\alias{FREEcvOutput}
\title{FORMAT OUTPUT FROM INTERNAL CROSS VALIDATION FUNCTIONS}
\description{Internal use only: formats the output of \code{FREEcvInternal}.}
\usage{
FREEcvOutput(x)
}
\arguments{
  \item{x}{List output from \code{FREEcvInternal}}
}
\value{
\item{observed}{observed response data}
\item{predicted}{predicted values for holdout data}
}
\author{
Jian Yen <jdl.yen@gmail.com>
}
\examples{
  \dontrun{
    ## FIX THIS UP
  }
}