% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{free-package}
\alias{free-package}
\title{Regularized Estimating Equations}
\description{
Unified regularized estimating equation solver. Currently include one solver with the l1 penalty only. More solvers and penalties are under development.
}
